1125 consolidate search filters into one file
1125 add search filter handlers file
